Store settings derived from prelogin responses, login responses, and EnvChange events separately from ClientContext so that ClientContext can be made immutable.

Propagate the negotiated encryption setting to the NetworkTransport so that it can used to programmatically enable and disable encryption when using TDS 7.4.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces the storage of negotiated settings from prelogin and login responses to enhance the connection and authentication process.
- `handler_factory.rs`: Added `NegotiatedSettings` and updated `SessionSettings` to store various session parameters and encryption settings.
- `session_handler.rs`: Implemented methods to validate prelogin and login results and negotiate settings.
- `tds_connection_provider.rs` and `tds_connection.rs`: Updated to use `NegotiatedSettings` instead of `SessionSettings`.
- `network_transport.rs`: Added encryption handling and notification methods.
- `login.rs`: Modified `EnvChangeProperties` to use `Option` types for certain fields.

Related work items: #33011, #33431

+++ tds-x/src/token/parsers.rs
@@ -155,12 +155,8 @@ impl TokenParser for LoginAckTokenParser {
         let build_hi = reader.read_byte().await?;
         let build_low = reader.read_byte().await?;
 
-        let prog_version = Version::new(
-            major,
-            minor,
-            (((build_hi as u16) << 8) | build_low as u16) as u16,
-            0,
-        );
+        let prog_version =
+            Version::new(major, minor, ((build_hi as u16) << 8) | build_low as u16, 0);
         Ok(Tokens::from(LoginAckToken {
             interface_type: interface,
             tds_version,
